renamed the bridge data scripts and heal bridges elev using 3m cells center points
